removing point & polygon from C API
