+++ src/geometry/geometry.h
@@ -1,7 +1,5 @@
 #pragma once
 
-#include "point.h"
-#include "polygon.h"
 #include "rtdoc.h"
 #include "query_iterator.h"
 #include "rtree.h"
