+++ src/geometry/rtree.h
@@ -11,6 +11,11 @@ extern "C" {
 
 struct RTree;
 
+enum QueryType {
+  CONTAINS,
+  WITHIN,
+};
+
 NODISCARD struct RTree *RTree_New();
 void RTree_Free(struct RTree *rtree) NOEXCEPT;
 void RTree_Insert(struct RTree *rtree, struct RTDoc const *doc);
@@ -20,8 +25,7 @@ NODISCARD bool RTree_IsEmpty(struct RTree const *rtree) NOEXCEPT;
 void RTree_Clear(struct RTree *rtree) NOEXCEPT;
 NODISCARD struct RTDoc *RTree_Bounds(struct RTree const *rtree);
 
-NODISCARD struct QueryIterator *RTree_Query_Contains(struct RTree const *rtree, struct RTDoc const *query);
-NODISCARD struct QueryIterator *RTree_Query_Within(struct RTree const *rtree, struct RTDoc const *query);
+NODISCARD struct QueryIterator *RTree_Query(struct RTree const *rtree, struct RTDoc const *queryDoc, enum QueryType queryType);
 
 NODISCARD size_t RTree_MemUsage(struct RTree const *rtree);
 #ifdef __cplusplus
